+++ game/cards/dm05/mecha_thunder.go
@@ -31,7 +31,7 @@ func LaByleSeekerOfTheWinds(c *match.Card) {
 	c.ManaCost = 7
 	c.ManaRequirement = []string{civ.Light}
 
-	c.Use(fx.Creature, fx.Blocker, func(card *match.Card, ctx *match.Context) {
+	c.Use(fx.Creature, fx.Blocker(), func(card *match.Card, ctx *match.Context) {
 
 		if event, ok := ctx.Event.(*match.CreatureDestroyed); ok {