+++ game/cards/dm06/angel_command.go
@@ -18,29 +18,26 @@ func GarielElementalOfSunbeams(c *match.Card) {
 
 	spellcast := false
 
-	c.Use(fx.Creature, fx.Doublebreaker, func(card *match.Card, ctx *match.Context) {
+	c.Use(fx.Creature, fx.Doublebreaker,
+		fx.When(fx.IHaveCastASpell, func(card *match.Card, ctx *match.Context) { spellcast = true }),
+		func(card *match.Card, ctx *match.Context) {
 
-		if _, ok := ctx.Event.(*match.SpellCast); ok {
-			spellcast = true
-		}
-
-		// remove persistent effect when turn ends
-		_, ok := ctx.Event.(*match.EndStep)
-		if ok {
-			spellcast = false
-		}
-
-		if event, ok := ctx.Event.(*match.PlayCardEvent); ok {
-			if event.CardID == card.ID {
-				if !spellcast {
-					ctx.InterruptFlow()
-					ctx.Match.WarnPlayer(card.Player, "You can summon this creature only if you have cast a spell this round")
-					return
-				}
+			// reset spellcast flag at the end of the turn
+			if _, ok := ctx.Event.(*match.EndStep); ok {
+				spellcast = false
 			}
 
-		}
+			if event, ok := ctx.Event.(*match.PlayCardEvent); ok {
+				if event.CardID == card.ID {
+					if !spellcast {
+						ctx.InterruptFlow()
+						ctx.Match.WarnPlayer(card.Player, "You can summon this creature only if you have cast a spell this turn.")
+						return
+					}
+				}
 
-	})
+			}
+
+		})
 
 }